requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
mypy-protobuf==3.3.0
pillow==10.3.0
pip-autoremove==0.10.0
protobuf==5.26.1
setuptools==69.5.1
sliver-py==0.0.19
tk==0.1.0
ttkbootstrap==1.10.1
types-protobuf==5.26.0.20240422
